Added three files to come up with a correct guess for the game.
LanguageArray.java contains a list of objects for each programming language and its data. An array contains all the language objects and there is a method to retrieve that array.  

SelectLanguage.java randomly selects a programming language from the array provided by the LanguageArray class. 

ChoiceData.java stores the randomly selected object and contains getters to easily retrieve the information for the selected language.